backend/requirements.txt
Django
djangorestframework
django-filter
psycopg2-binary
python-dotenv
google-generativeai
django-cors-headers
